logger/requirements.txt
APScheduler==3.10.1
azure-core==1.26.4
azure-cosmos==4.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
pytz==2023.3
requests==2.30.0
six==1.16.0
typing_extensions==4.6.1
tzlocal==5.0.1
urllib3==2.0.2
